refactor(gui): extract session command tests to sibling file
Move the ~1022-line `#[cfg(test)] mod tests` out of
`primer-gui/src/commands/session.rs` into a new sibling
`commands/session/tests.rs`, using the idiomatic Rust-2018
`session.rs` + `session/tests.rs` subdirectory form — the same
zero-churn pattern as PRs #268/#269/#270/#271/#272.

- session.rs drops 1841 -> 818 lines (-1023); it now holds only
  production command handlers + a 2-line `#[cfg(test)] mod tests;`
  declaration. The only added parent line is `mod tests;`.
- Pure verbatim relocation: no behaviour change, no incidental fixes.
- As a descendant module, the tests still reach every private parent
  item via `use super::*` — zero visibility churn, no signature changes.

Verification (all green):
- cargo test -p primer-gui --lib commands::session: 28 passed
  (default) / 28 (--no-default-features) — regression guard unchanged.
- cargo test --workspace: 51 "test result: ok", 0 failures.
- cargo clippy -p primer-gui --all-targets [--no-default-features]
  -- -D warnings: clean in both feature states.
- cargo fmt --all -- --check: clean.

README/ROADMAP need no change (neither references session.rs).
